+++ jest.config.js
@@ -4,6 +4,7 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
  */
 
-module.exports = require('pc-nrfconnect-shared/config/jest.config')([
-    'serialport',
-]);
+module.exports =
+    require('@nordicsemiconductor/pc-nrfconnect-shared/config/jest.config')([
+        'serialport',
+    ]);