+++ src/components/DeviceSelector.tsx
@@ -6,8 +6,11 @@
 
 import React from 'react';
 import { useDispatch } from 'react-redux';
-import { DeviceTraits } from '@nordicsemiconductor/nrf-device-lib-js';
-import { DeviceSelector, logger } from 'pc-nrfconnect-shared';
+import {
+    DeviceSelector,
+    logger,
+} from '@nordicsemiconductor/pc-nrfconnect-shared';
+import { DeviceTraits } from '@nordicsemiconductor/pc-nrfconnect-shared/nrfutil';
 
 import { closeDevice, openDevice } from '../actions/deviceActions';
 /**