+++ src/components/Terminal/Main.tsx
@@ -6,7 +6,7 @@
 
 import React, { useCallback, useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { PaneProps } from 'pc-nrfconnect-shared';
+import { PaneProps } from '@nordicsemiconductor/pc-nrfconnect-shared';
 
 import {
     getClearOnSend,
@@ -20,7 +20,7 @@ import {
 } from '../../features/terminal/terminalSlice';
 import Terminal from './Terminal';
 
-const Main = ({ active }: PaneProps) => {
+export default ({ active }: PaneProps) => {
     const dispatch = useDispatch();
     const serialPort = useSelector(getSerialPort);
     const serialOptions = useSelector(getSerialOptions);
@@ -107,5 +107,3 @@ const Main = ({ active }: PaneProps) => {
         />
     );
 };
-
-export default Main;