+++ src/hooks/useFitAddon.ts
@@ -9,12 +9,10 @@ import { FitAddon } from 'xterm-addon-fit';
 
 const fitAddon = new FitAddon();
 
-const useFitAddon = (height = 0, width = 0, lineMode = false) => {
+export default (height = 0, width = 0, lineMode = false) => {
     useEffect(() => {
         if (width * height > 0) fitAddon.fit();
     }, [width, height, lineMode]);
 
     return fitAddon;
 };
-
-export default useFitAddon;